Create execution environment in base 'FetchSpec' class
